ticket-16 Consolidate Classes into single module
