+++ tests/test_severity.py
@@ -8,9 +8,9 @@ from mongoengine import ValidationError
 # Taken from: https://stackoverflow.com/questions/714063/importing-modules-from-parent-folder
 currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
 parentdir = os.path.dirname(currentdir)
-sys.path.insert(0, parentdir)
+sys.path.insert(0, os.path.join(parentdir, "source"))
 
-from source.severity import *
+from audit_template import Severity, SeverityEnum
 
 
 class SeverityTests(unittest.TestCase):
